+++ src/Web/ViewModels/BasketItemViewModel.cs
@@ -2,8 +2,8 @@
 {
     public class BasketItemViewModel
     {
-        public int Id { get; set; }
-        public int CatalogItemId { get; set; }
+        public long Id { get; set; }
+        public long CatalogItemId { get; set; }
         public string ProductName { get; set; }
         public decimal UnitPrice { get; set; }
         public decimal OldUnitPrice { get; set; }
